Week 12 Shooter Game Build
Week 12 Build, completed task below

TASK 1: As the first task, have the "Lives: " text on the screen update to show the current lives value of the player.

TASK 2: As the second task, have a coin appear randomly on the screen. You can find a coin sprite in the assets. Have the coin stay on the screen for a couple of seconds and then destroy itself. If the player picks up the coin, they should earn "1" score. Alternatively, you can also have the coin move on the screen from one side to the other like the enemies.